Moved Skills to its own folder and fixed associated paths
